test-requirements.txt
pytest==8.2.2
mock==5.1.0
pytest-cov==5.0.0
tox==4.15.1
pytest-asyncio==0.23.7
httpx==0.27.0
coverage==7.5.3
ruff==0.5.1
pyyaml==6.0.1
